+++ NEWS.md
@@ -1,4 +1,6 @@
 # tern 0.7.4.9000
+* Added Design for MNG01 plot.
+* Added `stat_mean_sd` ggplot helper function to create mean +/- standard deviation interval as error bars.
 * Added new safety helper function h_stack_by_baskets to stack SMQ and/or CQ records from ADAE data set.
 * Added `remove_single` argument to `h_append_grade_groups`, which allows to control whether the elements of one-element grade groups will be kept in the output or removed.
 * Re-implemented the `h_append_grade_groups` function to improve its flexibility, robustness and clearness as well as to make sure the result is ordered according to the order of `grade_groups`.